I remove excess code.
