+++ components/GameScreen/GameScreenHeader.tsx
@@ -1,0 +1,9 @@
+const GameScreenHeader = () => (
+  <div className="flex flex-col items-center justify-center mt-10">
+    <h1 className="text-2xl font-bold text-center">Increase streams up to 10x.</h1>
+    <h1 className="text-2xl font-bold text-center">Turn any song into a video game.</h1>
+    <h1 className="text-2xl font-bold text-center">Link content, select game, sync streams.</h1>
+  </div>
+)
+
+export default GameScreenHeader
