--- components/GameScreen/components/MediaControls/index.tsx
@@ -1,3 +1,0 @@
-import MediaControls from "./MediaControls"
-
-export default MediaControls
