--- components/MintButton/index.ts
@@ -1,3 +1,0 @@
-import MintButton from "./MintButton"
-
-export default MintButton
