--- components/NavBar/NavBar.tsx
@@ -1,9 +1,0 @@
-import { ConnectButton } from "@rainbow-me/rainbowkit"
-
-const NavBar = () => (
-  <div className="flex justify-end mr-3">
-    <ConnectButton />
-  </div>
-)
-
-export default NavBar
