--- components/PopupModal/index.tsx
@@ -1,3 +1,0 @@
-import PopupModal from "./PopupModal"
-
-export default PopupModal
