--- components/SVG/TokenRow.js
@@ -1,129 +1,0 @@
-import Image from "next/image"
-import { useState } from "react"
-import { toast } from "react-toastify"
-import { allChains, useAccount, useNetwork, useSigner, useSwitchNetwork } from "wagmi"
-import { stake, unstake } from "../../lib/stake"
-
-const TokenRow = ({
-  tokenId,
-  nftContract,
-  staked,
-  image,
-  onSuccess,
-  setPendingTxStep,
-  className,
-}) => {
-  const { data: signer } = useSigner()
-  const { chain } = useNetwork()
-  const { address: account } = useAccount()
-  const { switchNetwork } = useSwitchNetwork()
-  const [hovering, setHovering] = useState(false)
-
-  const handleClick = async () => {
-    if (!signer) {
-      toast.error("please connect wallet")
-      return
-    }
-    if (chain.id !== Number(process.env.NEXT_PUBLIC_CHAIN_ID)) {
-      await switchNetwork(Number(process.env.NEXT_PUBLIC_CHAIN_ID))
-      const myChain = allChains.find(
-        (blockchain) => blockchain.id === process.env.NEXT_PUBLIC_CHAIN_ID,
-      )
-      toast.error(`Please connect to ${myChain.name} and try again`)
-      return
-    }
-    if (staked) {
-      await unstake(nftContract, tokenId, onSuccess, setPendingTxStep)
-    } else {
-      await stake(nftContract, tokenId, nftContract, onSuccess, setPendingTxStep, account)
-    }
-    setPendingTxStep(0)
-  }
-
-  let buttonFill
-  if (staked) {
-    buttonFill = hovering ? "#FD0101" : "#00a6c6"
-  } else {
-    buttonFill = hovering ? "#FD0101" : "#FAF400"
-  }
-
-  let buttonText
-  if (staked) {
-    buttonText = hovering ? "UNSTAKE" : "STAKED"
-  } else {
-    buttonText = "STAKE"
-  }
-
-  let buttonTransform
-  if (staked) {
-    buttonTransform = hovering ? 1097.28 : 1147.28
-  } else {
-    buttonTransform = 1197.28
-  }
-
-  return (
-    <div className="mt-[-5] sm:mt-[-10]">
-      <div className="relative top-[15vh] sm:top-[20vh] left-20 sm:left-40 h-[100px] sm:h-[130px] w-[100px] sm:w-130px">
-        <Image src={image} layout="fill" objectFit="cover" alt="chillrx" />
-      </div>
-
-      <svg
-        id="Layer_1"
-        className={className}
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 2062.81 1066.21"
-      >
-        <g>
-          <path
-            style={{ fill: "#030303" }}
-            d="M1436.27,11.74c103.7,.52,207.38,4.07,311.1-.91,23.66-1.14,47.33-.88,70.99-1.67,31.65-1.06,63.3-2.17,94.94-4.08,7.6-.46,9.8,3.87,9.66,20.81-1.11,131.34-2.04,262.69-2.18,394.03-.1,91.73-5.78,182.96-5.74,274.6,.04,87.27-2.28,174.6,.1,261.82,.49,18.14-.44,32.46-11.94,28.5-2.61-.9-5.52,2.57-8.23,2.19-6.25-.87-9,3.3-7.59,17.37,.64,6.32,.27,13.16,.19,19.75-.13,11.63-1.65,16.94-8.55,17.65-113.73,11.69-227.49,9.2-341.24,10.37-73.07,.75-146.16,.92-219.22,4.08-234.57,10.15-469.13,12.77-703.64,6.86-194.88-4.91-389.78-4.27-584.67-6.34-8.79-.09-20.41,9.61-25.62-4.41-4.41-11.89-2.15-35.11-2.54-53.35-4.54-216.22-.28-432.64-.56-648.96-.11-85.08-.12-170.16-.88-255.21-.2-22.78,4.08-31.93,13.96-29.01,11.45,3.39,15.74-5.31,13.77-29.24-1.79-21.8,3.76-28.4,13.21-28.24,31.31,.52,62.68,3.11,93.94,0C225.3-.64,315.08-1.75,404.78,2.1c221.98,9.53,444.06,3.36,666.05,9.38,71.68,1.94,143.33,4.74,215.03,.03,1.39-.01,2.78-.03,4.17-.04,2.78-.02,5.56-.03,8.33-.05,45.97,.11,91.94,.22,137.91,.32ZM617.27,989.47c217.83-.68,422.06,5.06,626.36-4.23,96.01-4.37,192.04-4.12,288.07-7.4,61.24-2.09,122.5,.7,183.75-.59,60.51-1.27,121.02-5.08,181.53-7.3,7.45-.27,10.37-5.18,10.16-21.78-.66-52.74-.69-105.53-.86-158.31-.29-89.46,3.43-178.7,4.75-268.02,2.36-159.88,2.18-319.87,3.53-479.81,.15-18.28-1.81-24.79-10.63-22.31-29.43,8.28-59.06,4.05-88.6,6.6-50.06,4.31-100.21,4.37-150.31,4.24-210.9-.55-421.79-1.79-632.68-2.87-151.74-.78-303.47-1.08-455.2-2.79-89.76-1.01-179.48-5.92-269.26-6.75-79.67-.74-159.25,10.37-238.88,8.29-7.31-.19-14.63,.03-21.91-.86-7.55-.93-10.89,3.05-10.66,21.68,.88,71.81,1.23,143.68,1.04,215.53-.38,140.87-1.11,281.74-1.98,422.6-.58,93.84,.42,187.56,2.42,281.24,.23,10.63,.49,16.77,7.3,16.68,59.51-.85,119.02-1.01,178.53-.67,135.71,.78,271.44-.09,393.54,6.84Z"
-          />
-          <path
-            style={{ fill: "#fefefe" }}
-            d="M617.27,989.47c-122.1-6.93-257.83-6.06-393.54-6.84-59.51-.34-119.02-.19-178.53,.67-6.81,.1-7.07-6.04-7.3-16.68-2-93.68-3-187.4-2.42-281.24,.87-140.87,1.6-281.74,1.98-422.6,.19-71.85-.16-143.72-1.04-215.53-.23-18.63,3.11-22.6,10.66-21.68,7.29,.89,14.61,.67,21.91,.86,79.64,2.08,159.21-9.03,238.88-8.29,89.78,.83,179.5,5.75,269.26,6.75,151.73,1.7,303.47,2.01,455.2,2.79,210.89,1.08,421.79,2.32,632.68,2.87,50.1,.13,100.25,.08,150.31-4.24,29.53-2.54,59.17,1.68,88.6-6.6,8.83-2.48,10.79,4.03,10.63,22.31-1.34,159.94-1.17,319.93-3.53,479.81-1.32,89.32-5.04,178.56-4.75,268.02,.17,52.78,.21,105.57,.86,158.31,.21,16.6-2.72,21.51-10.16,21.78-60.51,2.22-121.02,6.03-181.53,7.3-61.25,1.29-122.5-1.51-183.75,.59-96.03,3.28-192.06,3.04-288.07,7.4-204.3,9.29-408.53,3.55-626.36,4.23Z"
-          />
-        </g>
-        <g
-          onClick={handleClick}
-          onMouseEnter={() => setHovering(true)}
-          onMouseLeave={() => setHovering(false)}
-        >
-          <g>
-            <path
-              style={{ fill: "#040402" }}
-              d="M1055.08,480.9c.27-5.3,.59-9.87,.71-14.43,.12-4.39,1.68-5.57,7.07-5.65,28.96-.43,57.79,1.99,86.69,2.45,29.48,.47,58.97,.17,88.46,.09,17.28-.04,34.56-.44,51.84-.46,54.47-.04,108.94-.05,163.41,.08,16.9,.04,33.8,.51,50.7,.89,34.21,.78,68.34-.72,102.51-1.72,42.08-1.24,84.13-.65,126.16,1.3,2.77,.13,5.59,.01,8.34-.28,2.55-.27,3.96,.52,4.55,2.44,.22,.73,.28,1.51,.31,2.27,1.38,37.48,3.18,74.95,3.93,112.44,.38,19.41-1.08,38.86-1.8,58.28-.32,8.41-.88,16.82-1.33,25.23-1.25,23.55-2.42,47.1-3.82,70.64-.26,4.37-1.38,8.71-2.3,13.04-.5,2.33-2.71,3.06-5.4,3.24-7.67,.52-15.33,1.12-23.49,1.72-.17,.99-.55,2.14-.54,3.3,.05,3.83,.32,7.65,.35,11.47,.05,5.44-2.5,7.6-9,7.61-8.45,0-16.91-.16-25.36-.13-52.78,.13-105.56,.05-158.33,.53-48.47,.44-96.91-.12-145.34-1.58-24.4-.73-48.81-1.23-73.2-1.92-31.16-.87-62.31-2.16-93.47-2.62-25.34-.38-50.7,.28-76.05,.31-15.2,.01-30.41-.15-45.61-.51-18.96-.46-37.9-1.17-56.85-1.84-7.47-.27-9.73-1.97-9.85-8.11-.12-5.93,.58-11.86,.68-17.8,.56-30,.99-60,1.57-90,.7-36.12,1.5-72.24,2.3-108.37,.09-4.13,.7-8.27,.59-12.39-.29-10.55-.88-21.1-1.3-31.64-.15-3.67-.4-7.35-.18-11.01,.36-6.05,1.49-6.86,8.79-6.92,6.01-.05,12.02,.03,18.03,.05,1.85,0,3.7,0,6.27,0h-.01Zm678.02,264.37c.52-3.58,.81-6.27,1.34-8.92,.29-1.43,1.36-2.79,1.44-4.2,.89-16.67,1.6-33.34,2.48-50.01,1.47-27.83,3.6-55.65,4.36-83.5,.58-20.8-.34-41.63-.99-62.45-.58-18.36-1.69-36.71-2.6-55.06-.19-3.79-.55-7.56-.85-11.57-1.33-.09-2.42-.2-3.52-.23-13.14-.44-26.28-.82-39.41-1.33-39.07-1.51-78.12-.19-117.17,.7-22.9,.52-45.81,1.06-68.72,1.01-23.66-.05-47.32-1.15-70.98-1.15-64.24,0-128.49,.21-192.73,.6-50.53,.31-101.05,.72-151.51-2.27-10.22-.61-20.54-.09-31.38-.09-.66,12.94-1.38,25.61-1.93,38.27-.38,8.72-.43,17.45-.78,26.17-.79,19.58-1.69,39.16-2.51,58.75-.33,7.96-.33,15.93-.89,23.87-1.44,20.03-2.28,40.05-.66,60.12,1.55,19.09,2.62,38.2,3.95,57.3,.19,2.72,.65,5.43,1.01,8.33,1.79,.14,3.24,.34,4.69,.36,35.87,.46,71.75,.87,107.62,1.34,73.25,.96,146.5,1.77,219.74,3,41.5,.7,82.98,2.54,124.48,3.04,46.39,.55,92.8,.25,139.2,.15,17.46-.04,34.93-.42,52.39-.85,8.03-.2,16.04-.9,23.93-1.36h0Z"
-            />
-            <path
-              style={{ fill: buttonFill }}
-              d="M1733.1,745.26c-7.89,.46-15.9,1.16-23.93,1.36-17.46,.43-34.93,.82-52.39,.85-46.4,.09-92.8,.4-139.2-.15-41.5-.5-82.98-2.34-124.48-3.04-73.24-1.22-146.49-2.04-219.74-3-35.87-.47-71.75-.88-107.62-1.34-1.45-.02-2.89-.22-4.69-.36-.36-2.9-.82-5.61-1.01-8.33-1.33-19.1-2.4-38.21-3.95-57.3-1.63-20.07-.78-40.08,.66-60.12,.57-7.94,.56-15.91,.89-23.87,.82-19.58,1.72-39.16,2.51-58.75,.35-8.72,.41-17.45,.78-26.17,.55-12.67,1.26-25.33,1.93-38.27,10.83,0,21.15-.52,31.38,.09,50.46,2.99,100.98,2.57,151.51,2.27,64.24-.39,128.48-.6,192.73-.6,23.66,0,47.32,1.11,70.98,1.15,22.91,.04,45.81-.49,68.72-1.01,39.06-.89,78.1-2.22,117.17-.7,13.13,.51,26.27,.88,39.41,1.33,1.09,.04,2.18,.14,3.52,.23,.3,4.01,.66,7.79,.85,11.57,.92,18.35,2.03,36.7,2.6,55.06,.65,20.81,1.57,41.65,.99,62.45-.77,27.85-2.89,55.67-4.36,83.5-.88,16.67-1.59,33.34-2.48,50.01-.08,1.41-1.15,2.77-1.44,4.2-.53,2.66-.82,5.35-1.34,8.92h0Z"
-            />
-          </g>
-          <text
-            style={{
-              fontSize: "119.1px",
-              fontFamily: "MoreSugarRegular, 'More Sugar'",
-            }}
-            transform={`translate(${buttonTransform} 664.31)`}
-          >
-            <tspan x="0" y="0">
-              {buttonText}
-            </tspan>
-          </text>
-        </g>
-        <text
-          style={{
-            fontSize: "93.18px",
-            fontFamily: "MoreSugarThin, 'More Sugar'",
-          }}
-          transform="translate(1003.94 324.64)"
-        >
-          <tspan x="0" y="0">
-            TOKEN ID #{tokenId}
-          </tspan>
-        </text>
-      </svg>
-    </div>
-  )
-}
-
-export default TokenRow
