--- components/SkeletonCard/index.tsx
@@ -1,3 +1,0 @@
-import SkeletonCard from "./SkeletonCard"
-
-export default SkeletonCard
