--- components/StakingPage/index.ts
@@ -1,3 +1,0 @@
-import StakingPage from "./StakingPage"
-
-export default StakingPage
