+++ components/SyncButton/index.ts
@@ -1,0 +1,3 @@
+import SyncButton from "./SyncButton";
+
+export default SyncButton
