--- components/TextArea/index.tsx
@@ -1,3 +1,0 @@
-import TextArea from "./TextArea"
-
-export default TextArea
