--- components/TxModals/TxModal.js
@@ -1,140 +1,0 @@
-import SpinningPill from "./SpinningPill"
-
-const TxModal = ({
-  style,
-  stepOnePassed,
-  stepTwoPassed,
-  textOne,
-  textTwo,
-  translateOne,
-  translateTwo,
-}) => (
-  <div>
-    <SpinningPill />
-    <svg
-      id="Layer_1"
-      xmlns="http://www.w3.org/2000/svg"
-      style={style}
-      viewBox="0 0 1256.76 1114.73"
-    >
-      <g>
-        <path d="M485.45,0c38.32,.27,76.65,.37,114.97,.85,56.99,.72,113.97,1.99,170.96,2.46,24.31,.2,48.64-1.57,72.95-1.5,52.5,.15,104.95,1.95,157.27,6.86,43.27,4.06,85.84,11.48,127.15,25.12,15.96,5.27,31.56,11.67,47.16,17.96,13.04,5.26,23.79,14,33.57,23.92,3.86,3.91,7.22,8.31,11.09,12.21,12.77,12.86,17.79,28.68,17.68,46.35-.12,20.49,.3,41.03-1.07,61.44-1.85,27.57-.03,54.95,1.89,82.36,5.17,73.79,10.53,147.56,15.58,221.36,3.41,49.79,2.26,99.6-.25,149.38-3.45,68.39-7.51,136.76-10.39,205.18-1.42,33.61-.58,67.31-1.13,100.97-.38,23.24-4.17,45.78-16.15,66.25-.42,.72-.83,1.44-1.26,2.16-6.41,10.56-39.03,45.64-48.96,53.03-13.71,10.21-29.31,16.21-45.93,18.49-26.05,3.57-52.23,6.56-78.45,8.51-44.39,3.31-88.87,3.39-133.37,1.96-40.63-1.31-81.3-2.25-121.95-2.33-54.15-.11-108.31,.33-162.45,1.38-34.47,.67-68.91,2.9-103.37,4.38-36.12,1.55-72.25,2.91-108.37,4.6-37.49,1.75-74.96,2.11-112.42-.52-62.01-4.36-123.34-12.83-182.25-33.66-18.95-6.7-37.36-15.1-55.61-23.59-13.38-6.22-24.57-15.86-35.17-26.12-13.69-13.27-21.14-29.57-24.82-48.05-5.8-29.16-5.56-58.66-5.32-88.14,.5-63.83-1.32-127.6-3.55-191.39-1.75-50.1-1.02-100.29-1.59-150.44C1.33,504.28,.09,457.13,0,409.98c-.07-36.98,.61-73.98,1.98-110.94,1.36-36.8,3.64-73.57,6.16-110.31,1.96-28.73,3.32-57.58,10.14-85.75,3.96-16.36,11.67-30.37,25.36-40.73,2.2-1.66,4.09-3.94,5.58-6.28,5.47-8.62,12.59-15.48,21.2-20.85,26.12-16.29,54.6-25.77,85.24-27.39,34.09-1.8,68.25-2.66,102.39-3.33,66.3-1.3,132.61-2.23,198.92-3.21,9.48-.14,18.98,.33,28.46,.51l.02-1.7Zm-49.15,7.8c0,.55,0,1.1,0,1.65-58.48,.71-116.96,1.22-175.44,2.22-33.98,.58-67.96,1.5-101.9,3.09-30.49,1.43-58.88,10.41-84.97,26.51-14.27,8.81-23.55,21.45-26.96,37.52-3.62,17.06-6.57,34.41-7.95,51.78-2.97,37.36-5.24,74.8-6.89,112.25-1.89,42.94-3.64,85.91-3.91,128.88-.29,45.64,1.42,91.3,1.85,136.95,.35,36.82-.48,73.66,.25,110.47,.85,42.97,3.35,85.91,4.17,128.89,.91,47.64,.69,95.31,1.04,142.97,.17,23.34,.94,46.63,6.17,69.54,3.1,13.57,8.46,25.94,18.12,36.21,10.69,11.37,22.58,21.22,36.35,28.66,23.27,12.56,47.94,21.71,73.16,29.38,62.46,18.99,126.73,27.34,191.68,28.5,49.59,.89,99.26-1.63,148.88-3.18,30.3-.95,60.56-3.35,90.86-4.34,41.96-1.38,83.95-2.67,125.92-2.87,60.14-.29,120.31-.28,180.44,.91,57.99,1.15,115.95,3.22,173.83-1.8,23.22-2.02,46.48-4.19,69.49-7.77,32.83-5.1,57.73-22.47,73.12-52.3,9.06-17.56,11.77-36.7,12.05-56.19,.47-32.82-.24-65.68,1.14-98.46,2.7-63.76,6.65-127.46,9.82-191.2,2.61-52.43,4.96-104.91,1.42-157.37-4.04-59.84-8.92-119.62-13.1-179.45-2.29-32.74-4.82-65.51-5.53-98.3-.6-27.73,1.54-55.51,1.95-83.28,.29-19.1-.73-37.69-14.6-53.5-13.01-14.84-26.66-28.28-45-36.28-31.87-13.9-64.72-24.68-98.87-31.36-17.95-3.51-35.84-7.65-53.97-9.83-20.23-2.43-40.67-3.43-61.05-4.23-34.6-1.36-69.23-2.59-103.85-2.87-24.97-.2-49.95,1.93-74.92,1.74-39.13-.29-78.26-1.71-117.39-2.56-11.83-.26-23.66-.29-35.49-.33-63.31-.23-126.62-.42-189.93-.63Z" />
-        <path
-          style={{ fill: "#fff" }}
-          d="M436.3,7.8c63.31,.21,126.62,.4,189.93,.63,11.83,.04,23.67,.07,35.49,.33,39.13,.85,78.26,2.27,117.39,2.56,24.97,.19,49.95-1.95,74.92-1.74,34.62,.28,69.25,1.52,103.85,2.87,20.38,.8,40.82,1.8,61.05,4.23,18.12,2.18,36.02,6.32,53.97,9.83,34.14,6.68,67,17.46,98.87,31.36,18.34,8,31.99,21.44,45,36.28,13.87,15.82,14.89,34.4,14.6,53.5-.42,27.77-2.55,55.55-1.95,83.28,.71,32.79,3.24,65.56,5.53,98.3,4.18,59.83,9.06,119.61,13.1,179.45,3.55,52.47,1.19,104.94-1.42,157.37-3.17,63.74-7.12,127.44-9.82,191.2-1.39,32.78-.68,65.64-1.14,98.46-.28,19.49-2.99,38.64-12.05,56.19-15.39,29.83-40.29,47.2-73.12,52.3-23.01,3.58-46.27,5.75-69.49,7.77-57.89,5.03-115.85,2.95-173.83,1.8-60.13-1.19-120.29-1.21-180.44-.91-41.98,.21-83.96,1.49-125.92,2.87-30.3,.99-60.56,3.39-90.86,4.34-49.62,1.55-99.29,4.07-148.88,3.18-64.95-1.16-129.22-9.5-191.68-28.5-25.22-7.67-49.89-16.82-73.16-29.38-13.78-7.44-25.66-17.29-36.35-28.66-9.66-10.27-15.01-22.64-18.12-36.21-5.24-22.91-6-46.2-6.17-69.54-.35-47.66-.13-95.32-1.04-142.97-.82-42.97-3.32-85.91-4.17-128.89-.73-36.81,.1-73.65-.25-110.47-.43-45.65-2.15-91.31-1.85-136.95,.28-42.97,2.02-85.94,3.91-128.88,1.65-37.45,3.92-74.89,6.89-112.25,1.38-17.37,4.33-34.72,7.95-51.78,3.41-16.07,12.69-28.71,26.96-37.52,26.09-16.1,54.48-25.08,84.97-26.51,33.94-1.59,67.93-2.51,101.9-3.09,58.47-1,116.96-1.5,175.44-2.22,0-.55,0-1.1,0-1.65Z"
-        />
-      </g>
-      <text
-        style={{
-          fontSize: "95px",
-          fontFamily: "MoreSugarRegular, 'More Sugar'",
-        }}
-        transform="translate(413.35 224.07)"
-      >
-        <tspan x="0" y="0">
-          PENDING{" "}
-        </tspan>
-        <tspan x="-179.79" y="114">
-          TRANSACTION...
-        </tspan>
-      </text>
-      <text
-        style={{
-          fontSize: "65px",
-          fontFamily: "MoreSugarRegular, 'More Sugar'",
-        }}
-        transform={`translate(${translateOne} 832.57)`}
-      >
-        <tspan x="0" y="0">
-          {textOne}
-        </tspan>
-      </text>
-      <text
-        style={{
-          fontSize: "65px",
-          fontFamily: "MoreSugarRegular, 'More Sugar'",
-        }}
-        transform={`translate(${translateTwo} 947.98)`}
-      >
-        <tspan x="0" y="0">
-          {textTwo}
-        </tspan>
-      </text>
-      {stepOnePassed && (
-        <g>
-          <path
-            style={{ fill: "#080404" }}
-            d="M1134.93,758.77c.61,1.66,.5,2.85-1.34,3.97-.73,.44-1.36,.93-1.69,1.84-.18,.49-.87,.82-1.37,1.17-5.84,4.18-11.75,8.26-17.52,12.55-6.83,5.09-13.02,10.93-19.32,16.65-4.43,4.02-9.12,7.76-13.62,11.7-6.33,5.54-12.51,11.25-17.97,17.68-5.44,6.41-10.58,13.06-14.88,20.31-.34,.58-.69,1.16-.96,1.78-.34,.77-.79,1.29-1.71,1.26-.18,0-.42,.21-.53,.39-.28,.44-.52,.9-.76,1.37-.57,1.09-.99,1.33-2.2,1.1-6.88-1.35-13.6-3.3-20.19-5.7-2.48-.9-4.96-1.8-7.44-2.7-.17-.06-.35-.12-.52-.19-1.9-.68-2.82-1.89-3.06-4.04-1.21-10.86-2.49-21.72-3.96-32.54-.85-6.2-2.02-12.36-3.21-18.5-.91-4.73-2.07-9.41-3.09-14.12-.09-.44-.14-.93-.03-1.35,.07-.27,.42-.57,.71-.66,.72-.24,1.48-.37,2.21-.53,0-.62-.07-1.2,.03-1.75,.05-.26,.42-.62,.69-.67,6.49-1.17,12.99-2.32,19.49-3.45,1.24-.22,2.49-.36,3.73-.55,.18-.03,.38-.04,.52-.13,1-.6,1.1-.56,1.51,.53,1.65,4.43,2.31,9.09,3.3,13.68,1.31,6.04,1.27,12.19,1.73,18.31,.44,5.87,1.21,11.72,1.84,17.58,.02,.2,.09,.4,.16,.67,3.18-3.83,6.14-7.68,9.41-11.25,5.78-6.35,11.75-12.53,17.64-18.78,9.56-10.16,19.66-19.76,30.3-28.78,8.62-7.31,18.03-13.41,28.4-17.96,.31-.14,.6-.3,.9-.46,.92-.49,1.7-.37,2.36,.46,2.93,3.68,5.29,7.69,7.03,12.05,1.2,2.99,2.3,6.02,3.41,9.04Zm-1.83,2.63c.67-.47,.96-.9,.67-1.74-1.74-5.19-3.59-10.34-6.32-15.1-1.25-2.17-2.72-4.22-4.11-6.36-.66,.31-1.07,.52-1.49,.7-7.53,3.33-14.61,7.44-21.22,12.35-9,6.68-17.18,14.29-25.17,22.11-4.98,4.88-9.81,9.9-14.59,14.98-5.36,5.7-10.66,11.47-15.83,17.34-3.22,3.65-6.16,7.55-9.23,11.33-.4,.5-.82,.38-1.13-.06-.19-.26-.27-.62-.33-.94-.29-1.85-.65-3.69-.83-5.55-.58-6.17-1.09-12.35-1.61-18.53-.23-2.72-.26-5.47-.64-8.17-.55-3.79-1.29-7.56-2.05-11.32-.5-2.49-1.2-4.93-1.81-7.42-1.75,.27-3.3,.47-4.84,.74-5.93,1.02-11.85,2.06-17.78,3.08-.77,.13-.99,.4-.8,1.23,1.15,5.09,2.35,10.17,3.32,15.29,1.16,6.12,2.21,12.27,3.11,18.43,.91,6.26,1.58,12.55,2.34,18.83,.52,4.3,1.04,8.6,1.48,12.9,.08,.81,.33,1.33,1.05,1.65,.37,.16,.7,.4,1.07,.53,2.22,.82,4.45,1.6,6.66,2.43,6.51,2.44,13.16,4.39,19.96,5.84,1.73,.37,1.74,.37,2.61-1.19,3.87-6.91,8.73-13.14,13.71-19.25,7.37-9.03,15.95-16.85,24.67-24.52,4.17-3.67,8.58-7.08,12.66-10.84,5.94-5.48,11.89-10.95,18.36-15.79,5.95-4.45,12.06-8.68,18.12-12.98Z"
-          />
-          <path
-            style={{ fill: "#fd0000" }}
-            d="M1133.77,759.65c.28,.84,0,1.27-.67,1.74-6.06,4.3-12.17,8.53-18.12,12.98-6.47,4.84-12.42,10.32-18.36,15.79-4.08,3.76-8.49,7.17-12.66,10.84-8.72,7.67-17.3,15.49-24.67,24.52-4.98,6.11-9.84,12.33-13.71,19.25-.87,1.56-.88,1.56-2.61,1.19-6.79-1.46-13.45-3.4-19.96-5.84-2.21-.83-4.44-1.61-6.66-2.43-.37-.14-.7-.37-1.07-.53-.72-.32-.96-.84-1.05-1.65-.44-4.3-.96-8.6-1.48-12.9-.76-6.28-1.43-12.57-2.34-18.83-.9-6.16-1.95-12.31-3.11-18.43-.97-5.12-2.17-10.2-3.32-15.29-.19-.83,.03-1.1,.8-1.23,5.93-1.02,11.85-2.06,17.78-3.08,1.54-.26,3.08-.47,4.84-.74,.61,2.49,1.31,4.94,1.81,7.42,.76,3.75,1.5,7.53,2.05,11.32,.39,2.7,.42,5.44,.64,8.17,.52,6.18,1.03,12.36,1.61,18.53,.18,1.86,.53,3.7,.83,5.55,.05,.33,.14,.69,.33,.94,.31,.43,.73,.56,1.13,.06,3.07-3.79,6.01-7.68,9.23-11.33,5.18-5.87,10.48-11.64,15.83-17.34,4.77-5.08,9.61-10.1,14.59-14.98,7.99-7.82,16.17-15.44,25.17-22.11,6.61-4.91,13.69-9.02,21.22-12.35,.42-.18,.83-.39,1.49-.7,1.39,2.14,2.86,4.19,4.11,6.36,2.74,4.76,4.58,9.91,6.32,15.1Zm-10.27-15.14c0-.8-.64-1.45-1.45-1.46-.8,0-1.47,.63-1.5,1.42-.03,.79,.61,1.48,1.41,1.51,.81,.03,1.54-.66,1.54-1.48Zm-6.23,3.9c.69-.42,1.37-.85,1.41-1.79-.01-.79-.69-1.39-1.57-1.31-.45,.05-.92,.18-1.31,.4-5.76,3.24-11.43,6.61-16.16,11.31-.37,.37-.78,.71-1.08,1.13-.43,.61-.46,1.28,.05,1.87,.53,.62,1.22,.69,1.91,.3,.37-.21,.7-.5,1.01-.8,2.86-2.81,5.98-5.27,9.43-7.32,2.11-1.25,4.21-2.52,6.31-3.79Zm-75.26,92.23c0-.81-.7-1.48-1.56-1.5-.97-.02-1.86,.77-1.87,1.67-.01,.84,.67,1.51,1.54,1.49,1-.02,1.89-.8,1.9-1.66Zm-5.92-.64c.37-.62,.04-1.31-.55-1.67-.7-.43-1.46-.79-2.23-1.05-3.42-1.16-6.87-2.26-10.3-3.4-.7-.23-1.36-.58-2.06-.79-.83-.25-1.63,.13-1.92,.83-.3,.72-.01,1.51,.75,1.96,.37,.21,.77,.39,1.17,.53,3.66,1.23,7.32,2.45,10.98,3.67,.87,.29,1.75,.53,2.88,.88,.39-.28,1-.52,1.27-.96Zm-7.9-50.97c.04-.82-.59-1.52-1.37-1.53-.8-.01-1.48,.62-1.52,1.41-.04,.81,.59,1.5,1.38,1.53,.8,.03,1.47-.6,1.51-1.41Zm-.33-6.5c.08-.02,.16-.03,.24-.05-.85-4.05-1.7-8.09-2.56-12.13-.05-.23-.17-.46-.31-.66-.41-.59-.98-.83-1.68-.64-.67,.18-1.02,.67-1.05,1.35-.01,.36,.08,.72,.15,1.08,.53,2.61,1.08,5.22,1.59,7.84,.27,1.38,.42,2.79,.71,4.17,.16,.72,.61,1.32,1.47,1.26,.89-.06,1.24-.69,1.37-1.48,.04-.24,.04-.49,.06-.73Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M1122.05,743.06c.81,.01,1.46,.66,1.45,1.46,0,.82-.73,1.51-1.54,1.48-.8-.03-1.44-.72-1.41-1.51,.03-.79,.7-1.43,1.5-1.42Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M1118.68,746.63c-.03,.94-.71,1.37-1.41,1.79-2.1,1.27-4.2,2.54-6.31,3.79-3.45,2.04-6.57,4.51-9.43,7.32-.3,.3-.64,.59-1.01,.8-.69,.39-1.38,.31-1.91-.3-.51-.59-.48-1.26-.05-1.87,.3-.42,.71-.76,1.08-1.13,4.73-4.71,10.4-8.07,16.16-11.31,.39-.22,.86-.35,1.31-.4,.87-.09,1.56,.52,1.57,1.31Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M1040.45,839.15c.86,.02,1.56,.7,1.56,1.5,0,.86-.89,1.64-1.9,1.66-.86,.01-1.55-.65-1.54-1.49,.01-.9,.9-1.68,1.87-1.67Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M1035.54,838.35c.59,.37,.92,1.05,.55,1.67-.27,.44-.88,.68-1.27,.96-1.13-.34-2.01-.59-2.88-.88-3.66-1.22-7.32-2.44-10.98-3.67-.4-.14-.8-.31-1.17-.53-.77-.45-1.06-1.23-.75-1.96,.29-.7,1.09-1.09,1.92-.83,.7,.21,1.36,.56,2.06,.79,3.43,1.14,6.87,2.24,10.3,3.4,.78,.26,1.53,.62,2.23,1.05Z"
-          />
-          <path
-            style={{ fill: "#feefef" }}
-            d="M1026.82,787.52c.78,.01,1.41,.71,1.37,1.53-.03,.8-.71,1.44-1.51,1.41-.79-.03-1.42-.72-1.38-1.53,.03-.79,.72-1.42,1.52-1.41Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M1028.1,782.49c-.08,.02-.16,.04-.24,.05-.02,.24-.02,.49-.06,.73-.14,.79-.48,1.43-1.37,1.48-.86,.05-1.32-.54-1.47-1.26-.29-1.38-.44-2.79-.71-4.17-.51-2.62-1.06-5.23-1.59-7.84-.07-.36-.17-.72-.15-1.08,.03-.67,.38-1.17,1.05-1.35,.7-.19,1.27,.04,1.68,.64,.14,.2,.26,.43,.31,.66,.86,4.04,1.71,8.09,2.56,12.13Z"
-          />
-        </g>
-      )}
-
-      {stepTwoPassed && (
-        <g>
-          <path
-            style={{ fill: "#080404" }}
-            d="M884.35,957.57c2.21,.82,4.44,1.6,6.66,2.43,6.51,2.44,13.16,4.39,19.96,5.85,1.72,.37,1.74,.36,2.61-1.2,3.87-6.91,8.73-13.13,13.72-19.24,7.36-9.03,15.95-16.85,24.67-24.52,4.17-3.67,8.58-7.08,12.66-10.85,5.94-5.48,11.89-10.95,18.36-15.79,5.95-4.45,12.06-8.68,18.12-12.98,.67-.47,.95-.9,.67-1.74-1.74-5.19-3.59-10.34-6.32-15.1-1.25-2.17-2.72-4.22-4.11-6.36-.66,.31-1.07,.52-1.49,.71-7.53,3.33-14.61,7.44-21.22,12.35-9,6.68-17.19,14.29-25.18,22.11-4.98,4.88-9.81,9.9-14.58,14.98-5.36,5.7-10.66,11.47-15.84,17.34-3.22,3.65-6.16,7.55-9.23,11.34-.4,.5-.82,.37-1.13-.06-.19-.26-.28-.62-.33-.95-.3-1.85-.66-3.69-.83-5.55-.58-6.17-1.09-12.35-1.61-18.53-.23-2.72-.25-5.47-.64-8.17-.55-3.79-1.29-7.56-2.05-11.32-.5-2.49-1.2-4.93-1.81-7.42-1.75,.26-3.3,.47-4.84,.73-5.93,1.02-11.85,2.06-17.78,3.08-.77,.13-.99,.41-.8,1.23,1.15,5.09,2.35,10.17,3.32,15.29,1.16,6.12,2.21,12.26,3.11,18.43,.91,6.26,1.58,12.55,2.34,18.83,.52,4.3,1.04,8.59,1.48,12.9,.08,.81,.32,1.32,1.05,1.65,.36,.16,.7,.4,1.07,.53Zm118.58-78.93c.61,1.65,.5,2.85-1.34,3.97-.73,.44-1.36,.93-1.69,1.84-.18,.49-.87,.82-1.37,1.17-5.84,4.18-11.75,8.27-17.52,12.56-6.83,5.09-13.02,10.93-19.32,16.65-4.43,4.02-9.11,7.76-13.62,11.7-6.34,5.54-12.52,11.26-17.97,17.68-5.44,6.41-10.58,13.05-14.88,20.31-.35,.58-.7,1.16-.96,1.78-.34,.77-.79,1.29-1.71,1.26-.18,0-.42,.21-.53,.39-.28,.44-.52,.9-.76,1.37-.56,1.1-.99,1.34-2.19,1.1-6.88-1.35-13.6-3.29-20.19-5.7-2.48-.9-4.96-1.8-7.44-2.7l-.52-.18c-1.9-.68-2.82-1.89-3.06-4.04-1.21-10.86-2.49-21.72-3.96-32.54-.84-6.2-2.02-12.36-3.21-18.51-.91-4.73-2.07-9.41-3.09-14.12-.09-.43-.14-.92-.03-1.35,.07-.27,.42-.57,.7-.66,.72-.24,1.48-.37,2.21-.53,0-.61-.08-1.2,.03-1.74,.05-.27,.42-.62,.7-.67,6.49-1.18,12.99-2.32,19.48-3.46,1.24-.22,2.49-.36,3.74-.55,.18-.03,.37-.04,.52-.13,1-.6,1.1-.56,1.51,.53,1.65,4.43,2.31,9.09,3.3,13.68,1.31,6.04,1.27,12.19,1.73,18.31,.44,5.87,1.21,11.72,1.84,17.58,.02,.2,.1,.4,.17,.67,3.18-3.83,6.14-7.68,9.4-11.25,5.79-6.34,11.75-12.52,17.64-18.78,9.56-10.16,19.66-19.76,30.3-28.79,8.62-7.31,18.03-13.4,28.4-17.96,.31-.14,.6-.3,.9-.46,.92-.48,1.7-.37,2.36,.47,2.93,3.67,5.29,7.69,7.03,12.05,1.2,2.99,2.3,6.02,3.41,9.05Z"
-          />
-          <path
-            style={{ fill: "#fd0000" }}
-            d="M893.3,908.79c-.03,.81,.59,1.5,1.38,1.53,.8,.03,1.47-.6,1.51-1.41,.04-.82-.59-1.52-1.37-1.53-.8-.01-1.48,.62-1.52,1.41Zm2.1-23.9c.62,2.49,1.31,4.93,1.81,7.42,.76,3.76,1.5,7.53,2.05,11.32,.39,2.7,.42,5.45,.64,8.17,.52,6.18,1.04,12.35,1.61,18.53,.17,1.86,.53,3.7,.83,5.55,.05,.33,.14,.69,.33,.95,.31,.43,.73,.56,1.13,.06,3.07-3.79,6.01-7.69,9.23-11.34,5.18-5.87,10.48-11.64,15.84-17.34,4.77-5.08,9.6-10.1,14.58-14.98,7.99-7.82,16.18-15.44,25.18-22.11,6.61-4.91,13.69-9.02,21.22-12.35,.42-.19,.83-.39,1.49-.71,1.39,2.14,2.86,4.19,4.11,6.36,2.74,4.76,4.58,9.91,6.32,15.1,.28,.85,0,1.27-.67,1.74-6.06,4.3-12.18,8.53-18.12,12.98-6.47,4.84-12.42,10.32-18.36,15.79-4.08,3.76-8.49,7.18-12.66,10.85-8.72,7.67-17.3,15.49-24.67,24.52-4.99,6.11-9.84,12.33-13.72,19.24-.87,1.56-.88,1.56-2.61,1.2-6.8-1.46-13.45-3.41-19.96-5.85-2.21-.83-4.44-1.61-6.66-2.43-.37-.14-.71-.37-1.07-.53-.72-.32-.96-.83-1.05-1.65-.44-4.31-.96-8.6-1.48-12.9-.76-6.28-1.43-12.57-2.34-18.83-.9-6.17-1.95-12.31-3.11-18.43-.97-5.13-2.17-10.21-3.32-15.29-.19-.83,.03-1.1,.8-1.23,5.93-1.02,11.85-2.06,17.78-3.08,1.54-.26,3.08-.47,4.84-.73Zm96.1-20.51c0-.8-.64-1.45-1.45-1.46-.8,0-1.47,.63-1.5,1.42-.03,.79,.61,1.48,1.4,1.51,.82,.03,1.54-.66,1.54-1.48Zm-6.22,3.91c.69-.42,1.37-.85,1.4-1.79-.01-.79-.69-1.39-1.57-1.31-.45,.04-.92,.18-1.31,.39-5.76,3.24-11.43,6.61-16.16,11.32-.37,.37-.78,.71-1.08,1.13-.43,.6-.46,1.28,.05,1.87,.53,.61,1.22,.69,1.92,.3,.37-.21,.7-.5,1-.8,2.86-2.81,5.98-5.27,9.43-7.32,2.11-1.25,4.21-2.52,6.31-3.79Zm-75.27,92.23c0-.81-.7-1.48-1.56-1.5-.97-.02-1.86,.77-1.87,1.67-.01,.84,.67,1.5,1.53,1.49,1.01-.02,1.89-.79,1.9-1.66Zm-5.92-.64c.37-.62,.04-1.31-.55-1.67-.69-.43-1.46-.79-2.23-1.05-3.42-1.15-6.87-2.26-10.3-3.4-.7-.23-1.36-.58-2.06-.79-.83-.25-1.63,.13-1.92,.84-.3,.72-.01,1.51,.76,1.95,.37,.22,.77,.39,1.17,.53,3.66,1.23,7.32,2.46,10.98,3.67,.87,.29,1.75,.53,2.88,.88,.39-.28,1-.52,1.27-.96Zm-8.22-57.47c.08-.02,.16-.04,.23-.05-.85-4.05-1.7-8.09-2.56-12.14-.05-.23-.17-.46-.31-.66-.41-.6-.98-.83-1.68-.64-.67,.18-1.02,.67-1.05,1.35-.01,.36,.08,.72,.15,1.08,.53,2.61,1.08,5.22,1.59,7.84,.27,1.38,.42,2.79,.71,4.17,.16,.72,.61,1.32,1.47,1.26,.89-.05,1.24-.69,1.37-1.48,.04-.24,.04-.49,.06-.73Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M990.05,862.93c.81,.01,1.46,.66,1.45,1.46,0,.82-.73,1.51-1.54,1.48-.8-.03-1.44-.72-1.4-1.51,.03-.79,.7-1.43,1.5-1.42Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M986.68,866.5c-.03,.94-.71,1.37-1.4,1.79-2.1,1.27-4.2,2.54-6.31,3.79-3.45,2.04-6.57,4.51-9.43,7.32-.3,.3-.64,.59-1,.8-.69,.4-1.38,.31-1.92-.3-.51-.59-.48-1.27-.05-1.87,.3-.42,.71-.76,1.08-1.13,4.73-4.71,10.4-8.07,16.16-11.32,.39-.22,.86-.35,1.31-.39,.87-.09,1.56,.52,1.57,1.31Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M908.45,959.02c.86,.02,1.56,.69,1.56,1.5,0,.86-.89,1.64-1.9,1.66-.86,.01-1.55-.65-1.53-1.49,.01-.9,.89-1.69,1.87-1.67Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M903.54,958.22c.59,.36,.92,1.05,.55,1.67-.27,.45-.88,.68-1.27,.96-1.12-.35-2.01-.59-2.88-.88-3.66-1.22-7.32-2.44-10.98-3.67-.41-.14-.8-.31-1.17-.53-.77-.45-1.06-1.23-.76-1.95,.29-.71,1.09-1.09,1.92-.84,.7,.21,1.36,.56,2.06,.79,3.43,1.14,6.87,2.24,10.3,3.4,.78,.26,1.54,.62,2.23,1.05Z"
-          />
-          <path
-            style={{ fill: "#fff" }}
-            d="M896.1,902.36c-.08,.02-.16,.03-.23,.05-.02,.24-.02,.49-.06,.73-.13,.79-.48,1.43-1.37,1.48-.86,.06-1.32-.54-1.47-1.26-.3-1.37-.45-2.78-.71-4.17-.51-2.62-1.07-5.23-1.59-7.84-.07-.36-.17-.72-.15-1.08,.03-.68,.38-1.17,1.05-1.35,.7-.19,1.27,.04,1.68,.64,.14,.2,.26,.43,.31,.66,.86,4.05,1.71,8.09,2.56,12.14Z"
-          />
-          <path
-            style={{ fill: "#feefef" }}
-            d="M894.82,907.38c.78,.01,1.41,.71,1.37,1.53-.03,.81-.71,1.44-1.51,1.41-.79-.03-1.41-.72-1.38-1.53,.04-.79,.72-1.42,1.52-1.41Z"
-          />
-        </g>
-      )}
-    </svg>
-  </div>
-)
-
-export default TxModal
