--- components/TxModals/index.js
@@ -1,3 +1,0 @@
-import TxModal from "./TxModal"
-
-export default TxModal
