--- lib/consts.ts
@@ -1,22 +1,0 @@
-export const MUSIC_URLS = {
-  drums: [
-    "https://defient-music.s3.amazonaws.com/120_DRUMS_RNB_001-KATEE.wav",
-    "https://defient-music.s3.amazonaws.com/90_DRUMS_HIPHOP_001-SWEETS.wav",
-    "https://defient-music.s3.amazonaws.com/120_DRUMS_RNB_002-JOE.wav",
-  ],
-  vocal: [
-    "https://defient-music.s3.amazonaws.com/120_MISC_RNB_001-JEFF.wav",
-    "https://defient-music.s3.amazonaws.com/120_MUSIC_Bbmin_RNB_002-MCKENNA.wav",
-    "https://defient-music.s3.amazonaws.com/120_MUSIC_Bbmin_RNB_001-SERENA.wav",
-  ],
-  bass: [
-    "https://defient-music.s3.amazonaws.com/120_MUSIC_Bbmin_RNB_002-MCKENNA.wav",
-    "https://defient-music.s3.amazonaws.com/90_DRUMS_HIPHOP_002-SAMEER.wav",
-    "https://defient-music.s3.amazonaws.com/90_DRUMS_TRAP_001-MARION.wav",
-  ],
-  guitar: [
-    "https://defient-music.s3.amazonaws.com/90_MUSIC_Bbmin_HIPHOP_001-MICHAEL.wav",
-    "https://defient-music.s3.amazonaws.com/90_MISC_HIPHOP_001-SHANE.wav",
-    "https://defient-music.s3.amazonaws.com/90_MUSIC_Bbmin_HIPHOP_002-DOM.wav",
-  ],
-}
