--- lib/enums.ts
@@ -1,21 +1,0 @@
-export enum ChoiceOptions {
-  Undecided = "undecided",
-  PickYourCre8orType = "pick-your-cre8or-type",
-  FindYourCre8orType = "find-your-cre8or-type",
-}
-export enum RoadmapScreens {
-  Roadmap = "roadmap",
-  Trailer = "trailer",
-  Mysteries = "mysteries",
-  Storyline = "story",
-  AllowListChoice = "allow-list-choice",
-  ComingSoon = "roadmap",
-}
-export enum AllowListScreens {
-  AllowListChoice = "allow-list-choice",
-  PickYourCre8orType = "pick-your-cre8or-type",
-  FindYourCre8orType = "find-your-cre8or-type",
-  TypeformQuiz = "typeform-quiz",
-  Details = "details",
-  ComingSoon = "roadmap",
-}
