+++ lib/getTrackId.ts
@@ -1,0 +1,7 @@
+const getTrackId = (spotifyShareUrl: string) => {
+  const startIndex = spotifyShareUrl.indexOf("track/") + 6
+  const endIndex = spotifyShareUrl.indexOf("?")
+  return spotifyShareUrl.substring(startIndex, endIndex)
+}
+
+export default getTrackId
